Add Yes/No/All/Cancel confirmation on delete function.

+++ ftplugin/nerdtree.vim
@@ -5,7 +5,12 @@ execute "vnoremap <buffer> " . g:NERDTreeMapOpenInTab .    " :call <SID>ProcessS
 execute "vnoremap <buffer> dd :call <SID>ProcessSelection('Deleting', function('NERDTree_Delete'), 0)<CR>"
 
 function! NERDTree_Delete(node)
-    call a:node.delete()
+    if s:response < 3
+        let s:response = confirm("Delete " . a:node.path.str() . "?", "&Yes\n&No\n&All\n&Cancel")
+    endif
+    if s:response == 1 || s:response == 3
+        call a:node.delete()
+    endif
 endfunction
 
 function! NERDTree_Open(target, node)
@@ -15,6 +20,7 @@ function! NERDTree_Open(target, node)
 endfunction
 
 function! s:ProcessSelection(action, callback, closeWhenDone) range
+    let s:response = 0
     let curLine = a:firstline
     while curLine <= a:lastline
         call nerdtree#echo(a:action . " " . (curLine - a:firstline + 1) . " of " . (a:lastline - a:firstline + 1) . "...")
